fixed bug afflicting bar graph state, all graphs should be working now
