+++ src/components/graphs/bargraph/BarGraph.js
@@ -1,244 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import { Bar } from 'react-chartjs-2';
 
-const states = [
-  {
-    name: 'Alabama',
-    abbreviation: 'AL',
-  },
-  {
-    name: 'Alaska',
-    abbreviation: 'AK',
-  },
-  {
-    name: 'American Samoa',
-    abbreviation: 'AS',
-  },
-  {
-    name: 'Arizona',
-    abbreviation: 'AZ',
-  },
-  {
-    name: 'Arkansas',
-    abbreviation: 'AR',
-  },
-  {
-    name: 'California',
-    abbreviation: 'CA',
-  },
-  {
-    name: 'Colorado',
-    abbreviation: 'CO',
-  },
-  {
-    name: 'Connecticut',
-    abbreviation: 'CT',
-  },
-  {
-    name: 'Delaware',
-    abbreviation: 'DE',
-  },
-  {
-    name: 'District Of Columbia',
-    abbreviation: 'DC',
-  },
-  {
-    name: 'Federated States Of Micronesia',
-    abbreviation: 'FM',
-  },
-  {
-    name: 'Florida',
-    abbreviation: 'FL',
-  },
-  {
-    name: 'Georgia',
-    abbreviation: 'GA',
-  },
-  {
-    name: 'Guam',
-    abbreviation: 'GU',
-  },
-  {
-    name: 'Hawaii',
-    abbreviation: 'HI',
-  },
-  {
-    name: 'Idaho',
-    abbreviation: 'ID',
-  },
-  {
-    name: 'Illinois',
-    abbreviation: 'IL',
-  },
-  {
-    name: 'Indiana',
-    abbreviation: 'IN',
-  },
-  {
-    name: 'Iowa',
-    abbreviation: 'IA',
-  },
-  {
-    name: 'Kansas',
-    abbreviation: 'KS',
-  },
-  {
-    name: 'Kentucky',
-    abbreviation: 'KY',
-  },
-  {
-    name: 'Louisiana',
-    abbreviation: 'LA',
-  },
-  {
-    name: 'Maine',
-    abbreviation: 'ME',
-  },
-  {
-    name: 'Marshall Islands',
-    abbreviation: 'MH',
-  },
-  {
-    name: 'Maryland',
-    abbreviation: 'MD',
-  },
-  {
-    name: 'Massachusetts',
-    abbreviation: 'MA',
-  },
-  {
-    name: 'Michigan',
-    abbreviation: 'MI',
-  },
-  {
-    name: 'Minnesota',
-    abbreviation: 'MN',
-  },
-  {
-    name: 'Mississippi',
-    abbreviation: 'MS',
-  },
-  {
-    name: 'Missouri',
-    abbreviation: 'MO',
-  },
-  {
-    name: 'Montana',
-    abbreviation: 'MT',
-  },
-  {
-    name: 'Nebraska',
-    abbreviation: 'NE',
-  },
-  {
-    name: 'Nevada',
-    abbreviation: 'NV',
-  },
-  {
-    name: 'New Hampshire',
-    abbreviation: 'NH',
-  },
-  {
-    name: 'New Jersey',
-    abbreviation: 'NJ',
-  },
-  {
-    name: 'New Mexico',
-    abbreviation: 'NM',
-  },
-  {
-    name: 'New York',
-    abbreviation: 'NY',
-  },
-  {
-    name: 'North Carolina',
-    abbreviation: 'NC',
-  },
-  {
-    name: 'North Dakota',
-    abbreviation: 'ND',
-  },
-  {
-    name: 'Northern Mariana Islands',
-    abbreviation: 'MP',
-  },
-  {
-    name: 'Ohio',
-    abbreviation: 'OH',
-  },
-  {
-    name: 'Oklahoma',
-    abbreviation: 'OK',
-  },
-  {
-    name: 'Oregon',
-    abbreviation: 'OR',
-  },
-  {
-    name: 'Palau',
-    abbreviation: 'PW',
-  },
-  {
-    name: 'Pennsylvania',
-    abbreviation: 'PA',
-  },
-  {
-    name: 'Puerto Rico',
-    abbreviation: 'PR',
-  },
-  {
-    name: 'Rhode Island',
-    abbreviation: 'RI',
-  },
-  {
-    name: 'South Carolina',
-    abbreviation: 'SC',
-  },
-  {
-    name: 'South Dakota',
-    abbreviation: 'SD',
-  },
-  {
-    name: 'Tennessee',
-    abbreviation: 'TN',
-  },
-  {
-    name: 'Texas',
-    abbreviation: 'TX',
-  },
-  {
-    name: 'Utah',
-    abbreviation: 'UT',
-  },
-  {
-    name: 'Vermont',
-    abbreviation: 'VT',
-  },
-  {
-    name: 'Virgin Islands',
-    abbreviation: 'VI',
-  },
-  {
-    name: 'Virginia',
-    abbreviation: 'VA',
-  },
-  {
-    name: 'Washington',
-    abbreviation: 'WA',
-  },
-  {
-    name: 'West Virginia',
-    abbreviation: 'WV',
-  },
-  {
-    name: 'Wisconsin',
-    abbreviation: 'WI',
-  },
-  {
-    name: 'Wyoming',
-    abbreviation: 'WY',
-  },
-];
+import {
+  stateData,
+  createNumberOfIncidents,
+  createDataSet,
+} from '../assets/bargraphAssets';
 
 const def = {
   labels: ['Loading Data'],
@@ -255,68 +22,13 @@ const def = {
   ],
 };
 
-const stateData = {};
-
-for (let i = 0; i < states.length - 1; i++) {
-  stateData[states[i]['name']] = {
-    abbreviation: states[i]['abbreviation'],
-    count: 0,
-  };
-}
-
-stateData['Washington DC'] = {
-  abbreviation: 'DC',
-  count: 0,
-};
-
-stateData['Wyoming'] = {
-  abbreviation: 'WY',
-  count: 0,
-};
-
-stateData['Unknown'] = {
-  abbreviation: '???',
-  count: 0,
-};
-
-const createNumberOfIncidents = (data, stateData) => {
-  console.log('I am running');
-  data.forEach(incident => {
-    if (incident.state in stateData) {
-      stateData[incident.state]['count'] += 1;
-    } else {
-      stateData['Unknown']['count'] += 1;
-    }
-  });
-  console.log(stateData);
-  return stateData;
-};
-
-const createDataSet = data => {
-  const result = {
-    labels: Object.keys(stateData).map(key => key),
-    datasets: [
-      {
-        label: 'Number of incidents',
-        backgroundColor: 'rgb(103,183,220)',
-        borderColor: 'rgb(103,183,220)',
-        borderWidth: 1,
-        hoverBackgroundColor: 'rgba(255,99,132,0.4)',
-        hoverBorderColor: 'rgba(255,99,132,1)',
-        data: Object.keys(data).map(key => stateData[key]['count']),
-      },
-    ],
-  };
-  return result;
-};
-
-const BarGraph = ({ data }) => {
+const BarGraph = ({ count }) => {
   const [barData, setBarData] = useState(def);
   const [incidentCount, setIncidentCount] = useState(null);
 
   useEffect(() => {
-    setIncidentCount(createNumberOfIncidents(data, stateData));
-  }, []);
+    setIncidentCount(count);
+  }, [count]);
 
   useEffect(() => {
     if (incidentCount) {
